move TPF test help funcs to the same file
